requirements.txt
discord.py-self
httpx
asyncio
jsons